Put branch in working order
